refactor submission.page.ts
